add transactions and entry tests

+++ util/random.go
@@ -8,21 +8,21 @@ import (
 
 const alphabet = "abcdefghijklmnopqrstuvwxyz"
 
-func init(){
+func init() {
 	rand.Seed(time.Now().UnixNano())
 }
 
 // RandomInt generates a random intesget between min and max
-func RandomInt(min, max int64) int64{
-	return min + rand.Int63n(max - min + 1) 
+func RandomInt(min, max int64) int64 {
+	return min + rand.Int63n(max-min+1)
 }
 
 // RandomString generates a random string of length n
-func RandomString(n int) string{
+func RandomString(n int) string {
 	var sb strings.Builder
 	k := len(alphabet)
 
-	for i := 0; i < n; i++{
+	for i := 0; i < n; i++ {
 		c := alphabet[rand.Intn(k)]
 		sb.WriteByte(c)
 	}
@@ -31,24 +31,24 @@ func RandomString(n int) string{
 }
 
 // generate a random owner
-func RandomOwner() string{
+func RandomOwner() string {
 	return RandomString(6)
 }
 
 // generate a random amount of money
-func RandomMoney() int64{
+func RandomMoney() int64 {
 	return RandomInt(0, 1000)
 }
 
 // generate a random currency
-func RandomCurrency() string{
+func RandomCurrency() string {
 	currencies := []string{"USD", "GBP", "EUR", "CAD"}
 	n := len(currencies)
 	return currencies[rand.Intn(n)]
 }
 
 // generate random account IDs for FromAccount and ToAccount
-func RandomAccountIDs() (int,int){
-	fromAccountID, toAccountID := RandomInt(1,81), RandomInt(1,81)
-	return int(fromAccountID), int(toAccountID)
+func RandomAccountIDs() (int64, int64) {
+	fromAccountID, toAccountID := RandomInt(1, 81), RandomInt(1, 81)
+	return fromAccountID, toAccountID
 }
